refactor: navbar 수정, 미사용 파일 삭제
