+++ src/components/common/Navbar/Navbar.module.scss
@@ -1,4 +1,4 @@
-.NavbarV2 {
+.Navbar {
   display: flex;
   align-items: center;
   justify-content: space-between;
