--- src/components/ui/Navbar/Navbar.module.css
@@ -1,44 +1,0 @@
-@charset "UTF-8";
-.NavbarStory {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-.NavbarStory .Wrapper {
-  display: flex;
-  gap: 1rem;
-  justify-content: center;
-  align-items: center;
-}
-
-.Navbar {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-  height: 2.75rem;
-  padding: 0.75rem 1.25rem;
-  background-color: #fff; /* 배경색 (흰색) */
-}
-
-.Left,
-.Right {
-  display: flex;
-  align-items: center;
-  flex: 0 0 auto;
-}
-
-button {
-  background: none;
-  border: none;
-  cursor: pointer;
-}
-button .Logo {
-  height: 1.75rem;
-  vertical-align: bottom;
-  cursor: auto;
-}
-button img {
-  -o-object-fit: contain;
-     object-fit: contain;
-}
