--- src/components/ui/Navbar/Navbar.module.scss
@@ -1,47 +1,0 @@
-.NavbarStory {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-
-  .Wrapper {
-    display: flex;
-    gap: 1rem;
-    justify-content: center;
-    align-items: center;
-  }
-}
-
-.Navbar {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-  height: 2.75rem;
-  padding: 0.75rem 1.25rem;
-  background-color: #fff; /* 배경색 (흰색) */
-}
-
-.Left,
-.Right {
-  display: flex;
-  align-items: center;
-  flex: 0 0 auto;
-}
-
-button {
-  padding: 0;
-  margin: 0;
-  background: none;
-  border: none;
-  cursor: pointer;
-  height: 100%;
-  .Logo {
-    height: 1.75rem;
-    vertical-align: bottom;
-    cursor: auto;
-  }
-
-  img {
-    object-fit: contain;
-  }
-}
