+++ src/pages/HomePage.tsx
@@ -1,8 +1,8 @@
 import { useNavigate } from "react-router-dom";
 
+import Navbar from "@/components/common/Navbar/Navbar";
 import Home from "@/components/Home/Home";
 import Icon from "@/components/ui/Icon/Icon";
-import NavbarV2 from "@/components/ui/NavbarV2/NavbarV2";
 import Text from "@/components/ui/Text/Text";
 
 const HomePage = () => {
@@ -11,18 +11,19 @@ const HomePage = () => {
   const handleLogoClick = () => {
     navigate("/");
   };
+
   return (
     <>
-      <NavbarV2>
-        <NavbarV2.LeftButton onClick={handleLogoClick}>
+      <Navbar>
+        <Navbar.LeftButton onClick={handleLogoClick}>
           <Icon name="logo" />
-        </NavbarV2.LeftButton>
-        <NavbarV2.RightButton>
+        </Navbar.LeftButton>
+        <Navbar.RightButton>
           <Text variant="bodySm" color="secondary">
             앱 공유하기
           </Text>
-        </NavbarV2.RightButton>
-      </NavbarV2>
+        </Navbar.RightButton>
+      </Navbar>
 
       <Home />
     </>
